Redesign wallet. Again
More thorough isnpection shows that lots of places need updates
