--- interface/elements/msc-subcomponents/msc-main-status/msc-main-status.css
@@ -1,9 +1,0 @@
-
-#bottom_status_elements{
-}
-
-.mainStatus {
-  display: flex;
-  flex-wrap: nowrap;
-}
-
