+++ interface/styles/locales/en.json
@@ -1,9 +1,9 @@
 {
     "profileViewHtml_Btn_Copy_Account": "Copy to Clipboard",
-    "profileViewHtml_Table_Account": "Account",
+    "profileViewHtml_Table_Account": "Address",
     "profileViewHtml_Table_Actions": "Actions",
     "profileViewHtml_Table_Balance": "Balance",
-    "profileViewHtml_Table_Balance_in": "Balance in",
+    "profileViewHtml_Table_Balance_in": "View Balance in",
     "profileViewHtml_DD_MUSIC": "MUSIC",
     "profileViewHtml_DD_BTC": "BTC",
     "profileViewHtml_DD_USD": "USD",
@@ -30,9 +30,9 @@
     "profileViewHtml_Btn_NODES_LIST": "NODES LIST",
     "profileViewHtml_Btn_LOCATION": "SELECT LOCATION",
     "profileViewHtml_Btn_access_account": "Access your account",
-    "profileViewHtml_Table_ACCOUNT": "ACCOUNT",
+    "profileViewHtml_Table_ACCOUNT": "ADDRESS",
     "profileViewHtml_Table_ACTIONS": "ACTIONS",
-    "profileViewHtml_Table_BALANCE_IN": "BALANCE IN",
+    "profileViewHtml_Table_BALANCE_IN": "BALANCE",
     "profileViewHtml_Label_Message_To_Sign": "Message To Sign",
     "profileViewHtml_Label_Password": "Password",
     "profileViewHtml_Label_Re_Password": "Re Enter Password",
